Step2: Modularization
Implemented actual modules.
A module per 2d or 3rd level package (almost).
Currently, module can be external by being a mod at the same time. In
this case, module's methods will not be called by AE (except for IMC?).
This might be removed in the future.

Note: still a lot of work left to do on this stage. Logger needs to be
made non-static and build script is not complete yet.